feat: upgrade the aws-sdk to support the sqs json protocol (#440)
* chore: upgrade @aws-sdk/client-sqs to version which supports sqs json protocol

* fix: integration tests

* fix: remove v2 deprecated localstack envs

Localstack V2 deprecated the HOST_TMP_FOLDER and DATA_DIR envs. The DATA_DIR env was replaced with a PERSISTENCE env, but that is a paid feature of localstack now.

* fix: healthcheck for localstack verifies if sqs service is running

The localstack healthcheck endpoint now gives back a JSON object with a status for each service. This change makes it so that the healthcheck passes if the SQS service is in 'running' state.

* chore: fixing localstack version

* chore: updating types

* chore: updating test types

* chore: upgrading sqs-producer

* chore: update lock

Destroyed previously created lock and re ran to ensure it's BBC generated.

---------

Co-authored-by: Nicholas Griffin <nicholas.griffin@bbc.co.uk>